9480: adding to integration test, more report judges population [skip ci]
